Pearl_public/requirements.txt
aesara==2.9.4
aiofiles==24.1.0
aiohttp==3.11.11
aiosqlite==0.21.0
amdsmi==6.3.2
astunparse==1.6.3
attr==0.3.2
beautifulsoup4==4.13.3
blobfile==3.0.0
brotli==1.1.0
brotlicffi==1.1.0.0
carbon==1.1.10
chardet==5.2.0
click==8.1.8
conda==4.3.16
configobj==5.0.9
ConfigParser==7.1.0
contextlib2==21.6.0
coremltools==8.2
coverage==7.6.10
cryptography==44.0.0
ctypes_snappy==1.03
curio==1.6
Cython==3.0.11
deeplearning==0.0.1
dill==0.3.9
dl==0.1.0
docutils==0.21.2
duckduckgo_search==7.3.0
email_validator==2.2.0
etcd==2.0.8
eval_type_backport==0.2.2
exceptiongroup==1.2.2
expecttest==0.3.0
fadvise==6.0.1
fastparquet==2024.11.0
fbscribelogger==0.1.7
flint==0.0.1
gmpy2==2.2.1
graphviz==0.20.3
h2==4.2.0
HTMLParser==0.0.2
hypothesis==6.125.2
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.5
isal==1.7.1
jnius==1.1.0
js==1.0
kerchunk==0.2.8
keyring==25.6.0
lxml==5.3.0
lz4==4.4.3
lzmaffi==0.3.0
matchpy==0.5.5
monkeytype==23.3.0
moviepy==2.1.2
mtrand==0.1
multiset==3.2.0
nbformat==5.10.4
numarray==1.5.1
numba_rvsdg==0.0.5
Numeric==24.2
onnxscript==0.1.0
optree==0.14.0
outcome==1.3.0.post0
pandas==2.2.3
paramiko==3.5.1
pexpect==4.9.0
pickle5==0.0.12
Pillow==11.1.0
protobuf==5.29.3
psutil==6.1.1
py==1.11.0
pyarrow==19.0.0
pycosat==0.6.6
pydot==3.0.4
pydy==0.7.1
pyelftools==0.31
pygit2==1.17.0
pyglet==2.1.2
pygraphviz==1.14
PyInstaller==6.11.1
pyinstrument==5.0.1
pymc==5.20.0
pymc3==3.11.6
pynvml==12.0.0
pyodide==0.0.2
pyOpenSSL==25.0.0
pysat==3.2.1
python-dotenv==1.0.1
pytorch_lightning==2.5.0.post0
pytz==2025.1
pywin32==308
PyYAML==6.0.2
r2pipe==1.9.4
railroad==0.5.0
redis==5.2.1
requests_kerberos==0.15.0
rrdtool==0.1.16
safetensors==0.5.2
sage==0.0.0
scandir==1.10.0
schedule==1.2.2
scikit_learn==1.6.1
scipy==1.15.1
setuptools_scm==8.1.0
simplejson==3.19.3
slack_sdk==3.34.0
smbprotocol==1.15.0
socksio==1.0.0
Sphinx==8.1.3
spotipy==2.25.0
symengine==0.13.0
tabulate==0.9.0
textblob==0.19.0
theano==1.0.5
threadpoolctl==3.5.0
toml==0.10.2
torch_xla==2.6.0
torcharrow==0.1.0
torchaudio==2.6.0
torchrec==1.1.0
torchvision==0.21.0
trove_classifiers==2025.1.15.22
typeguard==4.4.1
ujson==5.10.0
urllib3_secure_extra==0.1.0
uvloop==0.21.0
websockets==14.2
xmlrpclib==1.0.1
xmlrunner==1.7.7
yarl==1.18.3
z3==0.2.0
zstandard==0.23.0
